+++ test/sample_files/opaque_struct_spec_module.ex
@@ -1,0 +1,5 @@
+defmodule Doctor.OpaqueStructSpecModule do
+  defstruct ~w(name arity)a
+
+  @opaque t :: %__MODULE__{}
+end